moved game to a new file
